Split sounds' app tests into multiple files
https://github.com/MTG/freesound/pull/1364
