+++ sounds/tests/test_sound.py
@@ -18,12 +18,12 @@
 #     See AUTHORS file.
 #
 
-import datetime
 import json
-import time
 import os
+import time
 
 import mock
+from bs4 import BeautifulSoup
 from django.conf import settings
 from django.contrib.auth.models import User, AnonymousUser
 from django.core import mail
@@ -32,143 +32,18 @@ from django.core.cache.backends import locmem
 from django.core.management import call_command
 from django.http import HttpRequest, HttpResponse
 from django.template import Context, Template
-from django.test import TestCase, Client, RequestFactory, override_settings
+from django.test import TestCase, RequestFactory, override_settings
 from django.urls import reverse
-from freezegun import freeze_time
 
 import accounts
+from accounts.models import EmailPreferenceType
 from comments.models import Comment
 from general.templatetags.filter_img import replace_img
-from sounds.models import Pack, Sound, SoundOfTheDay, License, DeletedSound, Flag
 from sounds.models import Download, PackDownload, PackDownloadSound, SoundAnalysis
-from sounds.views import get_sound_of_the_day_id
-from accounts.models import EmailPreferenceType
+from sounds.models import Pack, Sound, License, DeletedSound
+from utils.cache import get_template_cache_key
 from utils.encryption import encrypt
 from utils.test_helpers import create_user_and_sounds, override_analysis_path_with_temp_directory
-from utils.cache import get_template_cache_key
-
-from bs4 import BeautifulSoup
-
-
-class OldSoundLinksRedirectTestCase(TestCase):
-
-    fixtures = ['licenses', 'sounds']
-
-    def setUp(self):
-        self.sound = Sound.objects.all()[0]
-
-    def test_old_sound_link_redirect_ok(self):
-        # 301 permanent redirect, result exists
-        response = self.client.get(reverse('old-sound-page'), data={'id': self.sound.id})
-        self.assertEqual(response.status_code, 301)
-
-        # id is valid but has a space after it
-        response = self.client.get(reverse('old-sound-page'), data={'id': '%d ' % self.sound.id})
-        self.assertEqual(response.status_code, 301)
-
-    def test_old_sound_link_redirect_not_exists_id(self):
-        # 404 id does not exist
-        response = self.client.get(reverse('old-sound-page'), data={'id': 0}, follow=True)
-        self.assertEqual(response.status_code, 404)
-
-    def test_old_sound_link_redirect_invalid_id(self):
-        # 404 invalid id
-        response = self.client.get(reverse('old-sound-page'), data={'id': 'invalid_id'}, follow=True)
-        self.assertEqual(response.status_code, 404)
-
-
-class OldPackLinksRedirectTestCase(TestCase):
-
-    fixtures = ['packs']
-
-    def setUp(self):
-        self.client = Client()
-        self.pack = Pack.objects.all()[0]
-
-    def test_old_pack_link_redirect_ok(self):
-        response = self.client.get(reverse('old-pack-page'), data={'id': self.pack.id})
-        self.assertEqual(response.status_code, 301)
-
-        # id is valid but has a space after it
-        response = self.client.get(reverse('old-pack-page'), data={'id': '%d ' % self.pack.id})
-        self.assertEqual(response.status_code, 301)
-
-    def test_old_pack_link_redirect_not_exists_id(self):
-        response = self.client.get(reverse('old-pack-page'), data={'id': 0}, follow=True)
-        self.assertEqual(response.status_code, 404)
-
-    def test_old_pack_link_redirect_invalid_id(self):
-        response = self.client.get(reverse('old-pack-page'), data={'id': 'invalid_id'}, follow=True)
-        self.assertEqual(response.status_code, 404)
-
-
-class RandomSoundAndUploaderTestCase(TestCase):
-
-    fixtures = ['licenses', 'sounds']
-
-    def test_random_sound(self):
-        sound_obj = Sound.objects.random()
-        self.assertEqual(isinstance(sound_obj, Sound), True)
-
-
-class RandomSoundViewTestCase(TestCase):
-
-    fixtures = ['licenses']
-
-    @mock.patch('sounds.views.get_random_sound_from_solr')
-    def test_random_sound_view(self, random_sound):
-        """ Get a sound from solr and redirect to it. """
-        users, packs, sounds = create_user_and_sounds(num_sounds=1)
-        sound = sounds[0]
-
-        # We only use the ID field from solr
-        random_sound.return_value = {'id': sound.id}
-
-        response = self.client.get(reverse('sounds-random'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, '/people/testuser/sounds/{}/?random_browsing=true'.format(sound.id))
-
-    @mock.patch('sounds.views.get_random_sound_from_solr')
-    def test_random_sound_view_bad_solr(self, random_sound):
-        """ Solr may send us a sound id which no longer exists (index hasn't been updated).
-        In this case, use the database access """
-        users, packs, sounds = create_user_and_sounds(num_sounds=1)
-        sound = sounds[0]
-        # Update sound attributes to be selected by Sound.objects.random
-        sound.moderation_state = sound.processing_state = 'OK'
-        sound.is_explicit = False
-        sound.avg_rating = 8
-        sound.num_ratings = 5
-        sound.save()
-
-        # We only use the ID field from solr
-        random_sound.return_value = {'id': sound.id+100}
-
-        # Even though solr returns sound.id+100, we find we are redirected to the db sound, because
-        # we call Sound.objects.random
-        response = self.client.get(reverse('sounds-random'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, '/people/testuser/sounds/{}/?random_browsing=true'.format(sound.id))
-
-    @mock.patch('sounds.views.get_random_sound_from_solr')
-    def test_random_sound_view_no_solr(self, random_sound):
-        """ If solr is down, get a random sound from the database and redirect to it. """
-        users, packs, sounds = create_user_and_sounds(num_sounds=1)
-        sound = sounds[0]
-        # Update sound attributes to be selected by Sound.objects.random
-        sound.moderation_state = sound.processing_state = 'OK'
-        sound.is_explicit = False
-        sound.avg_rating = 8
-        sound.num_ratings = 5
-        sound.save()
-
-        # Returned if there is an issue accessing solr
-        random_sound.return_value = {}
-
-        # we find the sound due to Sound.objects.random
-        response = self.client.get(reverse('sounds-random'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, '/people/testuser/sounds/{}/?random_browsing=true'.format(sound.id))
 
 
 class CommentSoundsTestCase(TestCase):
@@ -549,159 +424,6 @@ class SoundViewsTestCase(TestCase):
         url = reverse('sound', args=['wrongusername', sound.id])
         resp = self.client.get(url)
         self.assertEqual(resp.status_code, 404)
-
-
-class RandomSoundTestCase(TestCase):
-    """ Test that sounds that don't fall under our criteria don't get selected
-        as a random sound."""
-
-    fixtures = ['licenses']
-
-    def _create_test_sound(self):
-        """Create a sound which is suitable for being chosen as a random sound"""
-        try:
-            user = User.objects.get(username="testuser")
-        except User.DoesNotExist:
-            user = None
-        user, packs, sounds = create_user_and_sounds(num_sounds=1, user=user)
-        sound = sounds[0]
-        sound.is_explicit = False
-        sound.moderation_state = 'OK'
-        sound.processing_state = 'OK'
-        sound.avg_rating = 8
-        sound.num_ratings = 5
-        sound.save()
-
-        return sound
-
-    def test_random_sound(self):
-        """Correctly selects a random sound"""
-        sound = self._create_test_sound()
-
-        random = Sound.objects.random()
-        self.assertEqual(random, sound)
-
-    def test_explict(self):
-        """Doesn't select a sound if it is marked as explicit"""
-        sound = self._create_test_sound()
-        sound.is_explicit = True
-        sound.save()
-
-        random = Sound.objects.random()
-        self.assertIsNone(random)
-
-    def test_not_processed(self):
-        """Doesn't select a sound if it isn't processed"""
-        sound = self._create_test_sound()
-        sound.processing_state = 'PE'
-        sound.save()
-
-        random = Sound.objects.random()
-        self.assertIsNone(random)
-
-        # or isn't moderated
-        sound = self._create_test_sound()
-        sound.moderation_state = 'PE'
-        sound.save()
-
-        random = Sound.objects.random()
-        self.assertIsNone(random)
-
-    def test_ratings(self):
-        """Doesn't select a sound if it doesn't have a high enough rating"""
-        sound = self._create_test_sound()
-        sound.avg_rating = 4
-        sound.save()
-
-        random = Sound.objects.random()
-        self.assertIsNone(random)
-
-    def test_flag(self):
-        """Doesn't select a sound if it's flagged"""
-        sound = self._create_test_sound()
-        sound.save()
-        Flag.objects.create(sound=sound, reporting_user=User.objects.all()[0], email="testemail@freesound.org",
-            reason_type="O", reason="Not a good sound")
-
-        random = Sound.objects.random()
-        self.assertIsNone(random)
-
-
-class SoundOfTheDayTestCase(TestCase):
-
-    fixtures = ['licenses', 'sounds', 'email_preference_type']
-
-    def setUp(self):
-        cache.clear()
-
-    def test_no_random_sound(self):
-        # If we have no sound, return None
-        random_sound_id = get_sound_of_the_day_id()
-        self.assertIsNone(random_sound_id)
-        # TODO: If we have some SoundOfTheDay objects, but not for day, don't return them
-
-    def test_random_sound(self):
-        sound = Sound.objects.get(id=19)
-        SoundOfTheDay.objects.create(sound=sound, date_display=datetime.date.today())
-        random_sound = get_sound_of_the_day_id()
-        self.assertEqual(isinstance(random_sound, int), True)
-
-    @freeze_time("2017-06-20")
-    def test_create_enough_new_sounds(self):
-        """ If we have some random sounds selected for the future, make sure
-        that we always have at least settings.NUMBER_OF_RANDOM_SOUNDS_IN_ADVANCE sounds
-        waiting in the future."""
-
-        sound_ids = []
-        user, packs, sounds = create_user_and_sounds(num_sounds=10)
-        for s in sounds:
-            sound_ids.append(s.id)
-
-        sound = Sound.objects.get(id=19)
-        SoundOfTheDay.objects.create(sound=sound, date_display=datetime.date(2017, 06, 20))
-        SoundOfTheDay.objects.create(sound=sound, date_display=datetime.date(2017, 06, 21))
-
-        call_command("create_random_sound")
-
-        sound_of_days = SoundOfTheDay.objects.count()
-        self.assertEqual(sound_of_days, 6)
-
-    def test_send_email_once(self):
-        """If we have a SoundOfTheDay, send the sound's user an email, but only once"""
-        sound = Sound.objects.get(id=19)
-        sotd = SoundOfTheDay.objects.create(sound=sound, date_display=datetime.date(2017, 06, 20))
-        sotd.notify_by_email()
-
-        self.assertEqual(len(mail.outbox), 1)
-        self.assertEqual(mail.outbox[0].subject, "[freesound] One of your sounds has been chosen as random sound of the day!")
-
-        # If we notify again, we don't send another email
-        sotd.notify_by_email()
-        self.assertEqual(len(mail.outbox), 1)
-
-    def test_user_disable_email_notifications(self):
-        """If the chosen Sound's user has disabled email notifications, don't send an email"""
-        sound = Sound.objects.get(id=19)
-
-        # Create email preference object for the email type (which will mean user does not want random sound of the
-        # day emails as it is enabled by default and the preference indicates user does not want it).
-        email_pref = accounts.models.EmailPreferenceType.objects.get(name="random_sound")
-        accounts.models.UserEmailSetting.objects.create(user=sound.user, email_type=email_pref)
-
-        sotd = SoundOfTheDay.objects.create(sound=sound, date_display=datetime.date(2017, 06, 20))
-        sotd.notify_by_email()
-
-        self.assertEqual(len(mail.outbox), 0)
-
-    @freeze_time("2017-06-20 10:30:00")
-    @mock.patch('django.core.cache.cache.set')
-    def test_expire_cache_at_end_of_day(self, cache_set):
-        """When we cache today's random sound, expire the cache at midnight today"""
-
-        sound = Sound.objects.get(id=19)
-        sotd = SoundOfTheDay.objects.create(sound=sound, date_display=datetime.date(2017, 06, 20))
-        sound_id = get_sound_of_the_day_id()
-        cache_set.assert_called_with("random_sound", 19, 48600)
 
 
 class DisplaySoundTemplatetagTestCase(TestCase):
@@ -1423,150 +1145,6 @@ class SoundAnalysisModel(TestCase):
         self.assertEqual(sa3.get_analysis(), None)
 
 
-class PublicSoundManagerTest(TestCase):
-    fixtures = ['licenses', 'sounds']
-
-    def test_public_sounds(self):
-        # Sounds.public manager only selects sounds which have
-        # processing state and moderation state set to OK
-        all_sounds = Sound.objects.all()
-        public_sounds = Sound.public.all()
-
-        self.assertEqual(len(all_sounds), len(public_sounds))
-
-        s1 = all_sounds[0]
-        s2 = all_sounds[1]
-        s1.processing_state = 'PE'
-        s1.save()
-        s2.moderation_state = 'PE'
-        s2.save()
-
-        all_sounds = Sound.objects.all()
-        public_sounds = Sound.public.all()
-
-        self.assertEqual(len(all_sounds), len(public_sounds) + 2)
-
-
-class SoundManagerQueryMethods(TestCase):
-
-    fixtures = ['licenses']
-
-    fields_to_check_bulk_query_id = ['username', 'id', 'type', 'user_id', 'original_filename', 'is_explicit',
-                                     'avg_rating', 'num_ratings', 'description', 'moderation_state', 'processing_state',
-                                     'processing_ongoing_state', 'similarity_state', 'created', 'num_downloads',
-                                     'num_comments', 'pack_id', 'duration', 'pack_name', 'license_id', 'license_name',
-                                     'license_deed_url', 'geotag_id', 'remixgroup_id', 'ac_analysis', 'tag_array']
-
-    fields_to_check_bulk_query_solr = ['username', 'user_id', 'id', 'type', 'original_filename', 'is_explicit',
-                                       'filesize', 'md5', 'channels', 'avg_rating', 'num_ratings', 'description',
-                                       'created', 'num_downloads', 'num_comments', 'duration', 'pack_id', 'geotag_id',
-                                       'bitrate', 'bitdepth', 'samplerate', 'pack_name', 'license_name', 'geotag_lat',
-                                       'geotag_lon', 'ac_analysis', 'is_remix', 'was_remixed', 'tag_array',
-                                       'comments_array']
-
-    def setUp(self):
-        user, packs, sounds = create_user_and_sounds(num_sounds=3, num_packs=1, tags="tag1 tag2 tag3")
-        self.sound_ids = [s.id for s in sounds]
-        self.user = user
-        self.pack = packs[0]
-
-    def test_bulk_query_id_num_queries(self):
-
-        # Check that all fields for each sound are retrieved with one query
-        with self.assertNumQueries(1):
-            for sound in Sound.objects.bulk_query_id(sound_ids=self.sound_ids):
-                for field in self.fields_to_check_bulk_query_id:
-                    self.assertTrue(hasattr(sound, field), True)
-
-    def test_bulk_query_id_field_contents(self):
-
-        # Check the contents of some fields are correct
-        for sound in Sound.objects.bulk_query_id(sound_ids=self.sound_ids):
-            self.assertEqual(Sound.objects.get(id=sound.id).user.username, sound.username)
-            self.assertEqual(Sound.objects.get(id=sound.id).original_filename, sound.original_filename)
-            self.assertEqual(Sound.objects.get(id=sound.id).pack_id, sound.pack_id)
-            self.assertEqual(Sound.objects.get(id=sound.id).license_id, sound.license_id)
-            self.assertItemsEqual(Sound.objects.get(id=sound.id).get_sound_tags(), sound.tag_array)
-
-    def test_bulk_query_solr_num_queries(self):
-
-        # Check that all fields for each sound are retrieved with one query
-        with self.assertNumQueries(1):
-            for sound in Sound.objects.bulk_query_solr(sound_ids=self.sound_ids):
-                for field in self.fields_to_check_bulk_query_solr:
-                    self.assertTrue(hasattr(sound, field), True)
-
-    def test_bulk_query_solr_field_contents(self):
-
-        # Check the contents of some fields are correct
-        for sound in Sound.objects.bulk_query_solr(sound_ids=self.sound_ids):
-            self.assertEqual(Sound.objects.get(id=sound.id).user.username, sound.username)
-            self.assertEqual(Sound.objects.get(id=sound.id).original_filename, sound.original_filename)
-            self.assertEqual(Sound.objects.get(id=sound.id).md5, sound.md5)
-            self.assertEqual(Sound.objects.get(id=sound.id).pack_id, sound.pack_id)
-            self.assertEqual(Sound.objects.get(id=sound.id).license_id, sound.license_id)
-            self.assertItemsEqual(Sound.objects.get(id=sound.id).get_sound_tags(), sound.tag_array)
-
-    def test_ordered_ids(self):
-
-        # This method is similar to SoundManager.bulk_query_id but returns the sounds in the same order as the
-        # the IDs in sound_ids. Here we only check that the sorting is correct. (the other things like returned
-        # sound fields being correct are already tested in previous tests).
-        with self.assertNumQueries(1):
-            for i, sound in enumerate(Sound.objects.ordered_ids(sound_ids=self.sound_ids)):
-                self.assertEqual(self.sound_ids[i], sound.id)
-
-    def test_bulk_sounds_for_user(self):
-
-        # This method uses SoundManager.bulk_query internally (also used by SoundManager.bulk_query_id) to retrieve
-        # sounds by a user. We only check that filtering by user works here (the other things like returned sound fields
-        # being correct are already tested in previous tests).
-
-        # Created sounds are not yet moderated and processed ok, so bulk_sounds_for_user should return no sounds
-        self.assertEqual(len(list(Sound.objects.bulk_sounds_for_user(user_id=self.user.id))), 0)
-
-        # Now we set user sounds to moderated and processed ok (and set user_sound_ids for later use)
-        user_sound_ids = []
-        for sound in Sound.objects.filter(user=self.user):
-            sound.moderation_state = 'OK'
-            sound.processing_state = 'OK'
-            sound.save()
-            user_sound_ids.append(sound.id)
-
-        # Check that now sounds returned by bulk_sounds_for_user are ok
-        user_sound_ids_bulk_query = []
-        with self.assertNumQueries(1):
-            for i, sound in enumerate(Sound.objects.bulk_sounds_for_user(user_id=self.user.id)):
-                self.assertEqual(self.user.id, sound.user_id)
-                user_sound_ids_bulk_query.append(sound.id)
-        self.assertItemsEqual(user_sound_ids, user_sound_ids_bulk_query)
-
-    def test_bulk_sounds_for_pack(self):
-
-        # This method uses SoundManager.bulk_query internally (also used by SoundManager.bulk_query_id) to retrieve
-        # sounds of a pack. We only check that filtering by pack works here (the other things like returned sound fields
-        # being correct are already tested in previous tests).
-
-        # Created sounds are not yet moderated and processed ok, so bulk_sounds_for_pack should return no sounds
-        self.assertEqual(len(list(Sound.objects.bulk_sounds_for_pack(pack_id=self.pack.id))), 0)
-
-        # Now we set user sounds to moderated and processed ok (and set pack_sound_ids for later use)
-        pack_sound_ids = []
-        for sound in Sound.objects.filter(pack=self.pack):
-            sound.moderation_state = 'OK'
-            sound.processing_state = 'OK'
-            sound.save()
-            pack_sound_ids.append(sound.id)
-
-        # Check that now sounds returned by bulk_sounds_for_pack are ok
-        pack_sound_ids_bulk_query = []
-        with self.assertNumQueries(1):
-            for i, sound in enumerate(Sound.objects.bulk_sounds_for_pack(pack_id=self.pack.id)):
-                self.assertEqual(self.user.id, sound.user_id)
-                pack_sound_ids_bulk_query.append(sound.id)
-        self.assertItemsEqual(pack_sound_ids, pack_sound_ids_bulk_query)
-
-
 class SoundEditDeletePermissionTestCase(TestCase):
     """Test that when editing and deleting sounds and packs only the user who owns
     them, or a specific admin can make the change"""
@@ -1644,4 +1222,3 @@ class SoundEditDeletePermissionTestCase(TestCase):
         self.client.force_login(self.other_user)
         resp = self.client.post(reverse('pack-delete', args=[self.pack.user.username, self.pack.id]))
         self.assertEqual(resp.status_code, 403)
-
